+++ src/fasta.h
@@ -19,7 +19,7 @@ size_t fasta_fread(FILE *fp, SeqRecord **records_ptr);
 size_t fasta_read(const char *path, SeqRecord **records_ptr);
 int fasta_fwrite(FILE *fp, SeqRecord *records, const size_t nrecords, const int maxlen);
 int fasta_write(const char *path, SeqRecord *records, const size_t nrecords, const int maxlen);
-void fasta_wrap_string(FILE *fp, const char *s, const int len, const int maxlen);
+void fasta_wrap_string(FILE *fp, const char *s, const size_t len, const int maxlen);
 char *fasta_get_id(const char *header);
 
 #endif // FASTA_H
