+++ src/str.h
@@ -13,7 +13,7 @@ typedef struct
 
 size_t str_count(const char *s, const char c);
 int str_is_in(const char **ss, unsigned int n, const char *t);
-int str_split(char ***fields_ptr, const char *s, const char d);
+ssize_t str_split(char ***fields_ptr, const char *s, const char d);
 void str_free_split(char **fields, const unsigned int n);
 
 #endif // STR_H
